third_party/WebKit/LayoutTests/pointer-lock/pointer-lock-api.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>
<script>
    description("Basic API existence test for Pointer Lock.")

    shouldBeDefined("document.onpointerlockchange");
    shouldBeDefined("document.onpointerlockerror");
    shouldBeDefined("document.exitPointerLock");
    shouldEvaluateTo("document.pointerLockElement", null);
    element = document.createElement('x');
    shouldBeDefined("element.requestPointerLock");
</script>
</body>
</html>
